requirements.txt
fastapi
metaflow
numpy
pandas
scikit-learn
sentence-transformers
uvicorn
seaborn
matplotlib
python-dotenv
groq